Fix GRPC server blocking interrupt of RSpec
The `GRPC::RpcServer` was run with `run_till_terminated_or_interrupted`
listening on the `QUIT`, `INT` and `TERM` signals. This however replaced
the original signal traps from RSpec, preventing its interruption after
the GRPC server was started.

To fix this, the signal traps are moved into the entrypoint, so they
are not registered during RSpec and RSpec continues to handle the
signals and shutting down the server.